fix: ci test failures + 2 pr review comments
Test failures fixed:
- MaxPool1DBackward: reverted to manual argmax scatter loop (CpuEngine
  TensorScatterAdd only supports 2D axis=0, argmax uses flat offsets)
- AdaptiveMaxPool2DBackward: same revert to manual scatter loop

PR review fixes:
- Vulkan ScatterMeanGlsl: indices buffer declared as int[] instead of
  float[] to correctly read integer indices without float reinterpretation
- To(DeviceInfo): already validates backend type match (repeat comment)

+++ src/AiDotNet.Tensors/Engines/DirectGpu/Vulkan/VulkanGlslKernels.cs
@@ -1273,7 +1273,7 @@ void main() {
 
     public static string ScatterMeanGlsl => Header + @"
 layout(set = 0, binding = 0) readonly buffer A { float a[]; };
-layout(set = 0, binding = 1) readonly buffer B { float bdata[]; };
+layout(set = 0, binding = 1) readonly buffer B { int bdata[]; };
 layout(set = 0, binding = 2) buffer C { uint c_bits[]; };
 layout(set = 0, binding = 3) buffer D { uint d_counts[]; };
 layout(push_constant) uniform Params { uint sourceSize; uint featureSize; };
